src/com/student/shared/dto/AssignmentDTO.java
package com.student.shared.dto;

public class AssignmentDTO extends DTO {
	
	private String name;
	private String description;
	private String dueDate;
	private boolean completed;
	private String repeat;
	private String taskID;
	private String calendarEventID;
	private String courseID;
	private boolean invalid;
	/** The ID of the assignment for which this instance is a repeated instance */
	private String head;
	
	public AssignmentDTO() {
		name = "";
		description = "";
		dueDate = null;
		completed = false;
		repeat = null;
		taskID = null;
		calendarEventID = null;
		courseID = null;
		invalid = false;
	}

	/**
	 * @return the name
	 */
	public String getName() {
		return name;
	}

	/**
	 * @param name the name to set
	 */
	public void setName(String name) {
		this.name = name;
	}

	/**
	 * @return the description
	 */
	public String getDescription() {
		return description;
	}

	/**
	 * @param description the description to set
	 */
	public void setDescription(String description) {
		this.description = description;
	}

	/**
	 * @return the dueDate
	 */
	public String getDueDate() {
		return dueDate;
	}

	/**
	 * @param dueDate the dueDate to set
	 */
	public void setDueDate(String dueDate) {
		this.dueDate = dueDate;
	}

	/**
	 * @return the completed
	 */
	public boolean isCompleted() {
		return completed;
	}

	/**
	 * @param completed the completed to set
	 */
	public void setCompleted(boolean completed) {
		this.completed = completed;
	}

	/**
	 * @return the repeat
	 */
	public String getRepeat() {
		return repeat;
	}

	/**
	 * @param repeat the repeat to set
	 */
	public void setRepeat(String repeat) {
		this.repeat = repeat;
	}

	/**
	 * @return the taskID
	 */
	public String getTaskID() {
		return taskID;
	}

	/**
	 * @param taskID the taskID to set
	 */
	public void setTaskID(String taskID) {
		this.taskID = taskID;
	}

	/**
	 * @return the calendarEventID
	 */
	public String getCalendarEventID() {
		return calendarEventID;
	}

	/**
	 * @param calendarEventID the calendarEventID to set
	 */
	public void setCalendarEventID(String calendarEventID) {
		this.calendarEventID = calendarEventID;
	}

	/**
	 * @return the courseID
	 */
	public String getCourseID() {
		return courseID;
	}

	/**
	 * @param courseID the courseID to set
	 */
	public void setCourseID(String courseID) {
		this.courseID = courseID;
	}

	/**
	 * @return the invalid
	 */
	public boolean isInvalid() {
		return invalid;
	}

	/**
	 * @param invalid the invalid to set
	 */
	public void setInvalid(boolean invalid) {
		this.invalid = invalid;
	}

	/**
	 * @return the head
	 */
	public String getHead() {
		return head;
	}

	/**
	 * @param head the head to set
	 */
	public void setHead(String head) {
		this.head = head;
	}
}
